fix: add typescript
