+++ index.html
@@ -5,7 +5,7 @@
     
     {{> head }} 
     
-    <link rel="stylesheet" type="text/css" href="/styles/themes/{{settings.theme}}.css" />
+    <link rel="stylesheet" type="text/css" href="/styles/themes/glitch.css" />
   </head>
   <body>
     <main>
