+++ layout/train.html
@@ -50,4 +50,4 @@
   </fluent-card>
   {{/if}}
 </div>
-<script type="module" src="../model/index.js"></script>
+<script type="module" src="../model/index.ts"></script>
