+++ types/index.ts
@@ -1,0 +1,3 @@
+export interface TensorData {
+    [key: string] : number[][]
+}
